master-decision-client/requirements.txt
six
grpcio
grpcio-tools
google
protobuf
requests
urllib3
charset_normalizer
idna
certifi
